requirements.txt
flask==2.2.3
flask-sqlalchemy==3.0.3
flask-login==0.6.2
opencv-python==4.7.0.72
opencv-contrib-python==4.7.0.72
numpy==1.24.2
pillow==9.4.0
pyyaml==6.0
python-dotenv==1.0.0
face-recognition==1.3.0
imutils==0.5.4
RPi.GPIO==0.7.1
picamera2==0.3.12
